fix: buffer conversion bug
